refactor: extract update params into IOU/UpdateMoneyRequestData.ts
Move getUpdateMoneyRequestParams and getUpdateTrackExpenseParams from
index.ts into new UpdateMoneyRequestData.ts module. Keep calculateDiffAmount
and getUpdatedMoneyRequestReportData in index.ts (used by other index.ts
functions). Add getRecentAttendees getter and export getPolicyTagsData,
maybeUpdateReportNameForFormulaTitle for the new module.

Update UpdateMoneyRequest.ts, Split.ts, MergeTransaction.ts,
TransactionUtils, IOUTest.ts, SplitDistanceMessageTest.ts.

Part of #72804
