+++ tests/actions/IOUTest.ts
@@ -9,7 +9,6 @@ import type {SearchQueryJSON, SearchStatus} from '@components/Search/types';
 import useOnyx from '@hooks/useOnyx';
 import {clearAllRelatedReportActionErrors} from '@libs/actions/ClearReportActionErrors';
 import {
-    calculateDiffAmount,
     clearBulkEditDraftTransaction,
     createDistanceRequest,
     handleNavigateAfterExpenseCreate,
@@ -34,6 +33,7 @@ import {
 import {putOnHold} from '@libs/actions/IOU/Hold';
 import {completeSplitBill, splitBill, startSplitBill, updateSplitTransactionsFromSplitExpensesFlow} from '@libs/actions/IOU/Split';
 import {requestMoney, trackExpense} from '@libs/actions/IOU/TrackExpense';
+import {calculateDiffAmount} from '@libs/actions/IOU/UpdateMoneyRequestData';
 import initOnyxDerivedValues from '@libs/actions/OnyxDerived';
 import {createWorkspace, generatePolicyID, setWorkspaceApprovalMode} from '@libs/actions/Policy/Policy';
 import {createNewReport, deleteReport, notifyNewAction} from '@libs/actions/Report';
